new commit

+++ src/usr_action.c
@@ -10,7 +10,7 @@ int game_roll(Game *g) /* ROLL：使用预置骰子移动 */
 
 int game_step(Game *g, int32_t steps) /* STEP：按指定步数移动 */
 {
-    int8_t last_position;
+    int32_t last_position;
 
     last_position = g->players[g->current_index].position;
     g->players[g->current_index].position += steps;
